+++ src/table/Header.js
@@ -15,19 +15,13 @@ export const THead = ({
   deletable,
   detailField,
   data,
-  tableOptions,
+  tableView,
   sortable,
-  filterable,
-  selectOptions,
   fromIndex,
   customProps
 }) => {
   const actions = getActions(schema, modelName)
   const onSort = R.path(['tableOptions', 'sort'], actions)
-  const onFilterChange = R.path(['tableOptions', 'filterChange'], actions)
-  const onFilterSubmit = R.path(['tableOptions', 'filterSubmit'], actions)
-  const onFilterRadio = R.path(['tableOptions', 'filterRadio'], actions)
-  const onMenuOpen = R.path(['input', 'onMenuOpen'], actions)
   return(
     <thead>
       <tr>
@@ -51,35 +45,23 @@ export const THead = ({
           }
 
           const isRelField = isRel(getField(schema, modelName, fieldName))
-          const filterInput = R.path(['filter', modelName, fieldName], tableOptions)
-          const sortKeyObj = R.path(['sort', modelName], tableOptions)
+          const sortKeyObj = R.path(['sort', modelName], tableView)
           return (
             <th key={idx} style={{ minWidth: '130px' }}>
               <Header
                 {...{
-                  schema,
                   modelName,
                   fieldName,
                   title: getFieldLabel({
                     schema, modelName, fieldName, data, customProps
                   }), // this is the actual 'data' list, not 'node'
-                  onFilterChange: (evt) => onFilterChange({
-                    modelName,
-                    ...evt
-                  }),
-                  onFilterSubmit,
-                  onFilterRadio,
                   onSort,
-                  onMenuOpen,
                   showSort: (
-                    tableOptions &&
+                    tableView &&
                     sortable &&
                     isSortable({schema, modelName, fieldName})
                   ) ? !isRelField : false,
-                  showFilter: isColFilterable({schema, modelName, fieldName, tableOptions, filterable}),
                   sortKeyObj,
-                  filterInput,
-                  selectOptions,
                 }}
               />
             </th>
@@ -90,20 +72,12 @@ export const THead = ({
 )}
 
 export const Header = ({
-  schema,
   modelName,
   fieldName,
   title,
-  onFilterChange,
-  onFilterSubmit,
-  onFilterRadio,
   onSort,
-  onMenuOpen,
   showSort,
-  showFilter,
   sortKeyObj,
-  filterInput,
-  selectOptions
 }) => (
   <div className='header'>
     <div className='title' >
@@ -113,7 +87,6 @@ export const Header = ({
       </a>
       <div className={'header-overflow'}>
         { showSort && <SortButton {...{ modelName, fieldName, onSort, sortKeyObj }} /> }
-        { showFilter && <FilterComp {... { fieldName, modelName, schema, onFilterChange, onFilterSubmit, onFilterRadio, onMenuOpen, filterInput, selectOptions }} /> }
       </div>
     </div>
   </div>